apisample15.html

<!DOCTYPE HTML>
<title>Example with a shared worker</title>
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>
<script>
test(function(t) {
  assert_true('SharedWorker' in self, "Browser should support SharedWorkers");
}, "Browser supports SharedWorkers");

fetch_tests_from_worker(new SharedWorker('apisample-worker.js', 'My shared worker'));

test(function(t) {
  assert_false(false, "False is false");
}, "Test running on main document. " +
"There should be an additional test named \'Test running on worker\'.");
</script>
